task3/requirements.txt
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
catboost==1.2.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configshell-fb==1.1.28
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
fonttools==4.46.0
future==0.18.2
graphviz==0.20.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.2
mccabe==0.7.0
mercurial==6.1.1
monotonic==1.6
more-itertools==8.10.0
multiprocess==0.70.16
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pandas==2.1.4
paramiko==2.9.3
pathos==0.3.2
pexpect==4.8.0
Pillow==9.0.1
plotly==5.18.0
pox==0.3.4
ppft==1.7.6.8
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pyan3 @ git+https://github.com/kuuurt/pyan.git@37404bb039bd9c5509b4aec8f61e360dfba50715
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
pyflakes==3.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-xlib==0.29
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
systemd-python==234
tenacity==8.2.3
threadpoolctl==3.2.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
urwid==2.1.2
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
